+++ branch-alpha/.artifact/screenshot_sh_install_download.svg.sh
@@ -9,5 +9,5 @@ printf '> #!/bin/sh
 
 echo "% Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                      Dload  Upload   Total   Spent    Left  Speed
-100  327k  100  327k    0     0   327k      0  0:00:01 --:--:--  0:00:01  327k"
+100  326k  100  326k    0     0   326k      0  0:00:01 --:--:--  0:00:01  326k"
 )